Implement readUser on Windows
