+++ solver/llbsolver/ops/file.go
@@ -167,7 +167,7 @@ func (f *fileOp) Exec(ctx context.Context, g session.Group, inputs []solver.Resu
 		inpRefs = append(inpRefs, workerRef.ImmutableRef)
 	}
 
-	fs := NewFileOpSolver(f.w, &file.Backend{}, f.refManager)
+	fs := NewFileOpSolver(f.w, &file.Backend{Executor: f.w.Executor()}, f.refManager)
 	outs, err := fs.Solve(ctx, inpRefs, f.op.Actions, g)
 	if err != nil {
 		return nil, err
